Fixes a bug where a promise was unhandled.
Fixes #10.

Rather than intermixing callbacks and promises, this opts to just use promises all the way through.

Giant caveat: This means that the `query` passed in to the constructor _must_ support promises.  I personally think this is the right/sane thing to do, but if you want to continue to support callbacks you will need to solve this bug in a different way (e.g., swallow the promise result of `query[queryMethod](...)`).