don't pass empty `newMethodCallback` function to ethjs-query call
This change to ethjs-query https://github.com/ethjs/ethjs-query/commit/dd1d1b29dc80a31da86704ae840ebdd4acf0e812 will not return a promise if any callback function is passed. This breaks promise returns for `.new()`, since this was always passing a function through to the ethjs-query `sendTransaction` call. With this commit, the callback function will only be passed to `sendTransaction` if a callback was defined.